refactor: extract shared API types to internal/api package
Move duplicated request/response types from internal/server and
internal/client into a shared internal/api/types.go package.

Both server and client now import from the shared package, reducing
duplication and ensuring consistency between API representations.

Closes #20
