+++ internal/client/client.go
@@ -7,6 +7,8 @@ import (
 	"io"
 	"net/http"
 	"time"
+
+	"github.com/rsclarke/oastrix/internal/api"
 )
 
 type HTTPClient interface {
@@ -41,69 +43,10 @@ func WithHTTPClient(httpClient HTTPClient) Option {
 	}
 }
 
-type CreateTokenRequest struct {
-	Label string `json:"label,omitempty"`
-}
 
-type CreateTokenResponse struct {
-	Token    string            `json:"token"`
-	Payloads map[string]string `json:"payloads"`
-}
 
-type InteractionResponse struct {
-	ID         int64                  `json:"id"`
-	Kind       string                 `json:"kind"`
-	OccurredAt string                 `json:"occurred_at"`
-	RemoteIP   string                 `json:"remote_ip"`
-	RemotePort int                    `json:"remote_port"`
-	TLS        bool                   `json:"tls"`
-	Summary    string                 `json:"summary"`
-	HTTP       *HTTPInteractionDetail `json:"http,omitempty"`
-	DNS        *DNSInteractionDetail  `json:"dns,omitempty"`
-}
-
-type HTTPInteractionDetail struct {
-	Method  string              `json:"method"`
-	Scheme  string              `json:"scheme"`
-	Host    string              `json:"host"`
-	Path    string              `json:"path"`
-	Query   string              `json:"query"`
-	Headers map[string][]string `json:"headers"`
-	Body    string              `json:"body"`
-}
-
-type DNSInteractionDetail struct {
-	QName    string `json:"qname"`
-	QType    int    `json:"qtype"`
-	QClass   int    `json:"qclass"`
-	RD       bool   `json:"rd"`
-	Opcode   int    `json:"opcode"`
-	DNSID    int    `json:"dns_id"`
-	Protocol string `json:"protocol"`
-}
-
-type GetInteractionsResponse struct {
-	Token        string                `json:"token"`
-	Interactions []InteractionResponse `json:"interactions"`
-}
-
-type ErrorResponse struct {
-	Error string `json:"error"`
-}
-
-type TokenInfo struct {
-	Token            string  `json:"token"`
-	Label            *string `json:"label"`
-	CreatedAt        string  `json:"created_at"`
-	InteractionCount int     `json:"interaction_count"`
-}
-
-type ListTokensResponse struct {
-	Tokens []TokenInfo `json:"tokens"`
-}
-
-func (c *Client) CreateToken(label string) (*CreateTokenResponse, error) {
-	reqBody := CreateTokenRequest{Label: label}
+func (c *Client) CreateToken(label string) (*api.CreateTokenResponse, error) {
+	reqBody := api.CreateTokenRequest{Label: label}
 	body, err := json.Marshal(reqBody)
 	if err != nil {
 		return nil, err
@@ -126,14 +69,14 @@ func (c *Client) CreateToken(label string) (*CreateTokenResponse, error) {
 		return nil, parseError(resp)
 	}
 
-	var result CreateTokenResponse
+	var result api.CreateTokenResponse
 	if err := json.NewDecoder(resp.Body).Decode(&result); err != nil {
 		return nil, err
 	}
 	return &result, nil
 }
 
-func (c *Client) GetInteractions(token string) (*GetInteractionsResponse, error) {
+func (c *Client) GetInteractions(token string) (*api.GetInteractionsResponse, error) {
 	req, err := http.NewRequest("GET", c.BaseURL+"/v1/tokens/"+token+"/interactions", nil)
 	if err != nil {
 		return nil, err
@@ -150,14 +93,14 @@ func (c *Client) GetInteractions(token string) (*GetInteractionsResponse, error)
 		return nil, parseError(resp)
 	}
 
-	var result GetInteractionsResponse
+	var result api.GetInteractionsResponse
 	if err := json.NewDecoder(resp.Body).Decode(&result); err != nil {
 		return nil, err
 	}
 	return &result, nil
 }
 
-func (c *Client) ListTokens() (*ListTokensResponse, error) {
+func (c *Client) ListTokens() (*api.ListTokensResponse, error) {
 	req, err := http.NewRequest("GET", c.BaseURL+"/v1/tokens", nil)
 	if err != nil {
 		return nil, err
@@ -174,7 +117,7 @@ func (c *Client) ListTokens() (*ListTokensResponse, error) {
 		return nil, parseError(resp)
 	}
 
-	var result ListTokensResponse
+	var result api.ListTokensResponse
 	if err := json.NewDecoder(resp.Body).Decode(&result); err != nil {
 		return nil, err
 	}
@@ -207,7 +150,7 @@ func parseError(resp *http.Response) error {
 		return fmt.Errorf("request failed with status %d", resp.StatusCode)
 	}
 
-	var errResp ErrorResponse
+	var errResp api.ErrorResponse
 	if err := json.Unmarshal(body, &errResp); err != nil || errResp.Error == "" {
 		return fmt.Errorf("request failed with status %d: %s", resp.StatusCode, string(body))
 	}
